+++ pyxem/tests/signals/test_indexation_results.py
@@ -25,6 +25,8 @@ from pyxem.signals.indexation_results import TemplateMatchingResults, VectorMatc
 from pyxem.signals.diffraction_vectors import DiffractionVectors
 from diffsims.libraries.vector_library import DiffractionVectorLibrary
 from diffsims.libraries.structure_library import StructureLibrary
+from diffsims.generators.diffraction_generator import DiffractionGenerator
+from diffsims.generators.library_generator import DiffractionLibraryGenerator
 from pyxem.utils.indexation_utils import OrientationResult
 from pyxem.generators.indexation_generator import TemplateIndexationGenerator
 from transforms3d.euler import euler2mat
@@ -36,11 +38,11 @@ def test_TemplateMatchingResults_to_crystal_map():
 
 def test_TemplateMatchingResults_plot_best_results_on_signal(diffraction_pattern,default_structure):
     """ Coverage testing """
-    edc = ds.DiffractionGenerator(300, 5e-2)
+    edc = DiffractionGenerator(300)
     half_side_length = 4
     rot_list = [[0,1,0],[1,0,0]]
 
-    diff_gen = ds.DiffractionLibraryGenerator(edc)
+    diff_gen = DiffractionLibraryGenerator(edc)
     struc_lib = StructureLibrary(["A"], [default_structure], [rot_list])
     library = diff_gen.get_diffraction_library(
         struc_lib,